Fix static issues in AudioD
:Release Notes:
Fix static issues related to parse recovery warnings and printf
arguments mismatch in AudioD

:Detailed Notes:
Addressed below static issues related to parse recovery warnings
and printf arguments mismatch in AudioD
CID 8480476 Uninitialized scalar field
CID 104274 Unchecked return value from library
CID 8480491 Resource leak in object
CID 6432450 Resource leak
CID 8481096 Printf arg type mismatch
CID 8481083 Printf arg type mismatch
CID 8480994 Printf arg type mismatch
CID 8480504 Printf arg type mismatch
CID 8480497 Printf arg type mismatch
CID 8480493 Printf arg type mismatch
CID 8481098 Printf arg count mismatch
CID 8481080 Printf arg count mismatch
CID 8480495 Printf arg count mismatch
CID 8481097 Parse recovery warning
CID 8481095 Parse recovery warning
CID 8481094 Parse recovery warning
CID 8481093 Parse recovery warning
CID 8481092 Parse recovery warning
CID 8481091 Parse recovery warning
CID 8481090 Parse recovery warning
CID 8481089 Parse recovery warning
CID 8481087 Parse recovery warning
CID 8481086 Parse recovery warning
CID 8481085 Parse recovery warning
CID 8481084 Parse recovery warning
CID 8481082 Parse recovery warning
CID 8481081 Parse recovery warning
CID 8481079 Parse recovery warning
CID 8481078 Parse recovery warning
CID 8481077 Parse recovery warning
CID 8481076 Parse recovery warning
CID 8481075 Parse recovery warning
CID 8481074 Parse recovery warning
CID 8481073 Parse recovery warning
CID 8481072 Parse recovery warning
CID 8481071 Parse recovery warning
CID 8481070 Parse recovery warning
CID 8481069 Printf arg type mismatch
CID 8481068 Parse recovery warning
CID 8481067 Parse recovery warning
CID 8481066 Parse recovery warning
CID 8481065 Parse recovery warning
CID 8481064 Parse recovery warning
CID 8481063 Parse recovery warning
CID 8481062 Parse recovery warning
CID 8481061 Parse recovery warning
CID 8481060 Parse recovery warning
CID 8481055 Printf arg type mismatch
CID 8481048 Printf arg type mismatch
CID 8481045 Printf arg type mismatch
CID 8481025 Printf arg type mismatch
CID 8481010 Printf arg type mismatch
CID 8481023 Parse warning

:Testing Performed:
Tested on RPI4 rebased on #6206 local gpvb build
Test case: QWA-4099, QWA-4098
Test Results are attached to PLAT-117896

:QA Notes:
NA

:Issues Addressed:
[PLAT-117896] Fix static analysis issues in AudioD for OSE

+++ src/audioMixer.cpp
@@ -193,7 +193,7 @@ void AudioMixer::callBackSinkStatus(const std::string& source, const std::string
 {
     PM_LOG_INFO(MSGID_AUDIO_MIXER, INIT_KVCOUNT,\
                 "callBackSinkStatus::source:%s sink:%s sinkId:%d sinkStatus:%d mixerType:%d",\
-                source.c_str(), sink.c_str(), audioSink, sinkStatus, mixerType);
+                source.c_str(), sink.c_str(), (int)audioSink, (int)sinkStatus, (int)mixerType);
     if (IsValidVirtualSink(audioSink))
     {
         if (utils::eSinkOpened == sinkStatus)
